% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_names.R
\name{anonymize_positions}
\alias{anonymize_positions}
\title{Anonymize Positions in Sentences}
\usage{
anonymize_positions(sentences_vector, positions_to_sanitize)
}
\arguments{
\item{sentences_vector}{A vector of sentences to be sanitized.}

\item{positions_to_sanitize}{A vector of position titles to be anonymized.}
}
\value{
A vector of sanitized sentences.
}
\description{
This function sanitizes occurrences of specified positions within a vector of sentences.
}
