% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anonymizing_whitelist}
\alias{anonymizing_whitelist}
\title{Whitelist of Terms}
\format{
A character vector with 380 elements.
\describe{
\item{Mona Lisa}{A term included in the whitelist.}
\item{Emily Carr}{A term included in the whitelist.}
\item{Prince George}{A term included in the whitelist.}
\item{...}{Additional terms included in the whitelist.}
}
}
\source{
Generated for example purposes
}
\usage{
anonymizing_whitelist
}
\description{
A character vector containing terms that are whitelisted for use in text data processing.
}
\details{
This vector is used to maintain specific terms in text data that should not be altered or anonymized.
}
\examples{
# Example usage
anonymizing_whitelist
# To access specific terms
anonymizing_whitelist[1:10]
}
\keyword{datasets}
