% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_and_merge_effect_sizes.R
\name{calculate_and_merge_effect_sizes}
\alias{calculate_and_merge_effect_sizes}
\title{This function takes an emmeans object and a model fit and returns a dataframe with effect sizes.}
\usage{
calculate_and_merge_effect_sizes(emmeans_obj, model)
}
\arguments{
\item{emmeans_obj}{An emmeans object created using data.frame(emmeans::emmeans(model)$contrasts)}

\item{model}{A model fit}
}
\value{
A dataframe with effect sizes
}
\description{
This function takes an emmeans object and a model fit and returns a dataframe with effect sizes.
}
