% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{capitalize_all_columns}
\alias{capitalize_all_columns}
\title{Capitalize All Character Columns}
\usage{
capitalize_all_columns(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
The original data frame with all character columns capitalized.
}
\description{
Converts the first letter of every character column in a data frame to uppercase.
}
