% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Capitalize First Letter}
\usage{
capitalize_first_letter(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with the first letter capitalized.
}
\description{
Capitalizes only the first letter of each string in the input vector.
}
