% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_basic.R
\docType{data}
\name{cnc_palette}
\alias{cnc_palette}
\title{CNC color palette}
\format{
An object of class \code{character} of length 15.
}
\usage{
cnc_palette
}
\description{
A character vector of hex color codes used as the cnc default palette.
}
\keyword{datasets}
