% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contraction_replacements}
\alias{contraction_replacements}
\title{Replacements for Contractions}
\format{
A named character vector with 14 elements:
\describe{
\item{weren't}{Replaced with "were not"}
\item{Weren't}{Replaced with "Were not"}
\item{don’t}{Replaced with "do not"}
\item{Don’t}{Replaced with "Do not"}
\item{don't}{Replaced with "do not"}
\item{Don't}{Replaced with "Do not"}
\item{haven't}{Replaced with "have not"}
\item{Haven't}{Replaced with "Have not"}
\item{weren't}{Replaced with "were not"}
\item{Weren't}{Replaced with "Were not"}
\item{don't}{Replaced with "do not"}
\item{Don't}{Replaced with "Do not"}
\item{haven't}{Replaced with "have not"}
\item{Haven't}{Replaced with "Have not"}
}
}
\source{
Generated for example purposes
}
\usage{
contraction_replacements
}
\description{
A named character vector containing contractions and their replacements.
}
\keyword{datasets}
