% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_long_dataframes.R
\name{create_long_dataframes}
\alias{create_long_dataframes}
\title{Create Long Dataframes and Save to Excel}
\usage{
create_long_dataframes(
  data,
  keywords,
  whitelist,
  filename = "data_long.xlsx",
  overwrite = FALSE
)
}
\arguments{
\item{data}{A dataframe to be pivoted to long format.}

\item{keywords}{A vector of keywords used to select columns for pivoting.}

\item{whitelist}{A vector of column names to be included in the final long dataframe along with the pivoted data.}

\item{filename}{The name of the output Excel file. Default is "data_long.xlsx".}

\item{overwrite}{A logical value indicating whether to overwrite the existing file if it already exists. Default is FALSE.}
}
\value{
None. The function saves the long dataframes as sheets in an Excel workbook.
}
\description{
This function pivots a dataframe to a long format for each specified keyword, processes the variable names,
and saves the resulting long dataframes as separate sheets in an Excel workbook.
}
\details{
The function performs the following steps:
\itemize{
\item Checks if the output file already exists. If it does and \code{overwrite} is FALSE, the function stops with an error message.
\item Creates a new Excel workbook.
\item Iterates over each keyword in \code{keywords}:
\item Pivots the dataframe to long format for columns containing the keyword.
\item Cleans up the variable names by removing the keyword, leading/trailing underscores, and double underscores.
\item Converts underscores to spaces and capitalizes the variable names.
\item Selects columns specified in the \code{whitelist} along with the cleaned variable names and their values.
\item Adds the long dataframe as a new sheet in the Excel workbook.
\item Saves the Excel workbook to the specified filename.
}
}
\examples{
\dontrun{
data <- data.frame(
  ID = 1:5,
  Age = c(23, 34, 45, 56, 67),
  Keyword_A_Score = c(10, 20, 30, 40, 50),
  Keyword_B_Score = c(15, 25, 35, 45, 55)
)
keywords <- c("Keyword_A", "Keyword_B")
whitelist <- c("ID", "Age")
create_long_dataframes(data, keywords, whitelist, filename = "data_long.xlsx", overwrite = TRUE)
}
}
