% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_curse_words.R
\name{curse_word_sanitization_function}
\alias{curse_word_sanitization_function}
\title{Curse Word Sanitization Function}
\description{
This script provides a function for sanitizing curse words in a given text by replacing them with a designated placeholder.
}
\section{Functions}{

\itemize{
\item \code{sanitize_curse_words} - Sanitizes curse words in a given text.
}
}

