% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{firstup}
\alias{firstup}
\title{Capitalize First Letter (Alternative)}
\usage{
firstup(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with the first letter in uppercase.
}
\description{
Capitalizes the first letter of each element in a character vector.
}
