% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_abbreviate.R
\name{fuzzy_abbreviate}
\alias{fuzzy_abbreviate}
\title{Apply Fuzzy Matching and Replacements to Abbreviate Text}
\usage{
fuzzy_abbreviate(text, corrections)
}
\arguments{
\item{text}{A character vector containing the text to be abbreviated.}

\item{corrections}{A data frame containing the original text and its corresponding replacement values.}
}
\value{
A character vector with abbreviated text.
}
\description{
This function applies fuzzy matching and replacements to abbreviate text based on a predefined corrections data frame.
}
