% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_abbreviate.R
\name{fuzzy_abbreviation_function}
\alias{fuzzy_abbreviation_function}
\title{Fuzzy Abbreviation Function}
\description{
This script provides a function for applying fuzzy matching and replacements to abbreviate text based on a predefined corrections data frame.
}
\section{Functions}{

\itemize{
\item \code{fuzzy_abbreviate} - Applies fuzzy matching and replacements to abbreviate text.
}
}

