% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_neutralize.R
\name{gender_neutralize}
\alias{gender_neutralize}
\title{Replace Gender-Specific Pronouns with Gender-Neutral Alternatives}
\usage{
gender_neutralize(sentence)
}
\arguments{
\item{sentence}{A character vector containing the text to be processed.}
}
\value{
A character vector with gender-specific pronouns replaced by gender-neutral alternatives.
}
\description{
This function replaces gender-specific pronouns (e.g., "he", "she") with gender-neutral alternatives (e.g., "they").
}
