% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary.R
\name{make_summary}
\alias{make_summary}
\title{This function generates a summary table with various statistics for the specified grouping variables and dependent variable.}
\usage{
make_summary(data, group_vars, summarization_var)
}
\arguments{
\item{data}{The targeted dataframe}

\item{group_vars}{a character string/vector of your grouping variables}

\item{summarization_var}{a character string of your targeted summary variable}
}
\value{
A dataframe with summary statistics
}
\description{
This function generates a summary table with various statistics for the specified grouping variables and dependent variable.
}
