% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_names.R
\name{max_typos}
\alias{max_typos}
\title{Determine Maximum Allowed Typos Based on Word Length}
\usage{
max_typos(word_length)
}
\arguments{
\item{word_length}{The length of the word.}
}
\value{
The maximum number of allowed typos.
}
\description{
This function determines the maximum number of allowed typos based on the length of the word.
}
