% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{mean_excluding_minus_99}
\alias{mean_excluding_minus_99}
\title{Mean Excluding -99}
\usage{
mean_excluding_minus_99(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The mean of the values in \code{x} that are not \code{-99}.
}
\description{
Calculates the mean of a numeric vector excluding values equal to -99.
}
