% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_emmeans_summary.R
\name{merge_emmeans_summary}
\alias{merge_emmeans_summary}
\title{This function takes summary data and a tidy emmeans object and adds emmeans-related columns to the summary data}
\usage{
merge_emmeans_summary(summary_data, emmean_obj)
}
\arguments{
\item{summary_data}{The  dataframe output of run_summary}

\item{emmean_obj}{The tidy dataframe}
}
\value{
A dataframe with summary statistics
}
\description{
This function takes summary data and a tidy emmeans object and adds emmeans-related columns to the summary data
}
