% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_many_emmeans_summary.R
\name{merge_many_emmeans_summary}
\alias{merge_many_emmeans_summary}
\title{Merge and Rename Columns in Multiple Dataframes}
\usage{
merge_many_emmeans_summary(x, y, grouping_vars)
}
\arguments{
\item{x}{A list of summary dataframes.}

\item{y}{Another list emmean dataframes.}

\item{grouping_vars}{A character vector specifying the grouping variables for merging.}
}
\value{
A list of dataframes with specified columns renamed.
}
\description{
Merges two lists of dataframes by specified grouping variables and renames
selected columns.
}
\examples{
\dontrun{
# Example usage
# This is an example, replace df1, df2, df3, df4 with your actual data frames
result <- merge_many_emmeans_summary(list(df1, df2), list(df3, df4), c("species", "sex"))
}

}
