% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_basic.R
\docType{data}
\name{nova_palette}
\alias{nova_palette}
\title{A clean default ggplot2 theme}
\format{
An object of class \code{character} of length 8.
}
\usage{
nova_palette
}
\value{
A ggplot2 theme object.
}
\description{
\code{theme_basic()} creates a clean plot theme inspired by \code{cowplot::theme_half_open()},
with bold axis/legend titles and markdown support for subtitles.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  theme_basic()
Nova color palette

A character vector of hex color codes used as the Nova default palette.

}
\keyword{datasets}
