% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{positions_to_sanitize}
\alias{positions_to_sanitize}
\title{Positions to Sanitize}
\format{
A character vector with 13 elements.
\describe{
\item{associate dean}{A position included in the list of positions to sanitize.}
\item{dean}{A position included in the list of positions to sanitize.}
\item{Vice President}{A position included in the list of positions to sanitize.}
\item{president}{A position included in the list of positions to sanitize.}
\item{department head}{A position included in the list of positions to sanitize.}
\item{head of the department}{A position included in the list of positions to sanitize.}
\item{program coordinator}{A position included in the list of positions to sanitize.}
\item{program head}{A position included in the list of positions to sanitize.}
\item{program chair}{A position included in the list of positions to sanitize.}
\item{area head}{A position included in the list of positions to sanitize.}
\item{area chair}{A position included in the list of positions to sanitize.}
\item{chair}{A position included in the list of positions to sanitize.}
\item{advisor}{A position included in the list of positions to sanitize.}
}
}
\source{
Generated for example purposes
}
\usage{
positions_to_sanitize
}
\description{
A character vector containing positions or titles that need to be sanitized in text data.
}
\details{
This vector is used to identify and sanitize specific job titles or positions in text data to ensure anonymity or consistency.
}
\examples{
# Example usage
positions_to_sanitize
# To access specific positions
positions_to_sanitize[1:5]
}
\keyword{datasets}
