% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_neutralize.R
\name{pronoun_neutralization_functions}
\alias{pronoun_neutralization_functions}
\title{Neutralize Pronouns}
\description{
This script provides functions for neutralizing pronouns in a given text, replacing gender-specific pronouns with gender-neutral alternatives.
It includes functions for context-aware replacement of "her" and handling verb conjugations, as well as general gender-neutralization of pronouns.
}
\section{Functions}{

\itemize{
\item \code{context_aware_her_replace} - Performs context-aware replacement of "her" and handles verb conjugations.
\item \code{gender_neutralize} - Replaces gender-specific pronouns with gender-neutral alternatives.
}
}

