% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{propagate_headers}
\alias{propagate_headers}
\title{Propagate Headers}
\usage{
propagate_headers(headers)
}
\arguments{
\item{headers}{A character vector of header values.}
}
\value{
A character vector with propagated header values.
}
\description{
Propagates non-empty header values from the first row to subsequent empty header cells.
}
