% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{recode_columns}
\alias{recode_columns}
\title{Recode Columns}
\usage{
recode_columns(df, columns, old_values, new_values)
}
\arguments{
\item{df}{A data frame.}

\item{columns}{A character vector specifying which columns to recode.}

\item{old_values}{A vector of old values to be replaced.}

\item{new_values}{A vector of new values to replace the old values.}
}
\value{
The data frame with recoded columns.
}
\description{
Recode values in specified columns of a data frame.
}
