% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_contractions.R
\name{replace_contractions}
\alias{replace_contractions}
\title{Replace Contractions in Text}
\usage{
replace_contractions(text, replacements)
}
\arguments{
\item{text}{A character vector containing the text to be processed.}

\item{replacements}{A named character vector specifying the contractions and their expanded forms.}
}
\value{
A character vector with contractions replaced by their expanded forms.
}
\description{
This function replaces contractions in text with their expanded forms. It also cleans special apostrophe characters before conducting replacements.
}
