% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{replace_strings}
\alias{replace_strings}
\title{Replace Strings in Variable Names}
\usage{
replace_strings(name, replacements)
}
\arguments{
\item{name}{A character string.}

\item{replacements}{A named character vector where names indicate the substrings to be replaced
and values provide the replacement strings.}
}
\value{
A character string with the replacements made.
}
\description{
Replaces substrings in a given name using a named vector of replacements.
}
