% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_pval_full.R
\name{report_pval_full}
\alias{report_pval_full}
\title{report p-value}
\usage{
report_pval_full(pval, italicize = TRUE)
}
\arguments{
\item{pval}{the p value}

\item{italicize}{Whether the symbols should be italicized}
}
\value{
A character string of p-value
}
\description{
report p-value
}
