% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tidy_anova_etaci.R
\name{report_tidy_anova_etaci}
\alias{report_tidy_anova_etaci}
\title{In-text and in-plot reporting of anova}
\usage{
report_tidy_anova_etaci(
  tidy_frame,
  term,
  effsize = TRUE,
  ci95 = TRUE,
  ci_lab = TRUE,
  teststat = TRUE,
  pval = TRUE
)
}
\arguments{
\item{tidy_frame}{Your tidy pairwise comparison result}

\item{term}{The predictor in your tidy ANOVA dataframe}

\item{effsize}{Display the effect size}

\item{ci95}{Display the 95\% CI}

\item{ci_lab}{Display the "95\% CI" label}

\item{teststat}{Display the F-score and degrees of freedom}

\item{pval}{Whether to include the p value}
}
\value{
A character string of stats
}
\description{
In-text and in-plot reporting of anova
}
