% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tidy_t.R
\name{report_tidy_t}
\alias{report_tidy_t}
\title{In-text and in-plot reporting of pairwise comparisons}
\usage{
report_tidy_t(
  tidy_frame,
  italicize = TRUE,
  ci = TRUE,
  ci_lab = TRUE,
  teststat = FALSE,
  point = TRUE,
  pval = TRUE,
  pval_comma = TRUE
)
}
\arguments{
\item{tidy_frame}{Your tidy pairwise comparison result}

\item{italicize}{Whether the symbols should be italicized}

\item{ci}{Whether to include confidence intervals}

\item{ci_lab}{Whether to include the label for confidence intervals}

\item{teststat}{Whether to include the test statistic}

\item{point}{Whether to include the point estimate for cohen's d}

\item{pval}{Whether to include the p value}

\item{pval_comma}{Whether to include a leading comma for p values}
}
\value{
A character string of stats
}
\description{
In-text and in-plot reporting of pairwise comparisons
}
