% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_many_anova.R
\name{run_many_anovas}
\alias{run_many_anovas}
\title{This function runs car::Anova on lm models, returning each result to a list}
\usage{
run_many_anovas(lm_results_list, type = NA)
}
\arguments{
\item{lm_results_list}{A list of lm model objects}

\item{type}{An optional character string specifying the type of ANOVA to be performed (default is "III")}
}
\value{
A list of ANOVA results
}
\description{
This function runs car::Anova on lm models, returning each result to a list
}
