% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_many_emmeans.R
\name{run_many_emmeans}
\alias{run_many_emmeans}
\title{Run emmeans on a list of lm results}
\usage{
run_many_emmeans(lm_results_list, emmeans_specs)
}
\arguments{
\item{lm_results_list}{A list of lm results.}

\item{emmeans_specs}{The emmeans specifications to be passed to the emmeans function.}
}
\value{
A list of emmeans results corresponding to each lm model in the input list.
}
\description{
This function takes a list of lm results and performs emmeans analysis on each
model based on the specified emmeans specifications.
}
