% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_many_lm.R
\name{run_many_lm}
\alias{run_many_lm}
\title{This function loops over multiple variables with run_anova, returning each one to a list}
\usage{
run_many_lm(data, independent_variable, dependent_variables)
}
\arguments{
\item{data}{The targeted dataframe}

\item{independent_variable}{character string of our independent variable}

\item{dependent_variables}{1+ character strings with numeric variable names}
}
\value{
A list of dataframes with ANOVA results
}
\description{
This function loops over multiple variables with run_anova, returning each one to a list
}
