% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_many_summaries.R
\name{run_many_summaries}
\alias{run_many_summaries}
\title{This function loops over multiple variables with run_summary, returning each one to a list}
\usage{
run_many_summaries(data, group_vars, summarization_vars)
}
\arguments{
\item{data}{The targeted dataframe}

\item{group_vars}{character strings of our grouping variables}

\item{summarization_vars}{1+ character strings with numeric variable names}
}
\value{
A list of dataframes with with summary statistics
}
\description{
This function loops over multiple variables with run_summary, returning each one to a list
}
