% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_names.R
\name{sanitize_courses_in_sentences}
\alias{sanitize_courses_in_sentences}
\title{Sanitize Courses in Sentences}
\usage{
sanitize_courses_in_sentences(sentences_vector, courses_to_sanitize)
}
\arguments{
\item{sentences_vector}{A vector of sentences to be sanitized.}

\item{courses_to_sanitize}{A vector of course names to be anonymized.}
}
\value{
A vector of sanitized sentences.
}
\description{
This function sanitizes occurrences of specified courses within a vector of sentences.
}
