% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_curse_words.R
\name{sanitize_curse_words}
\alias{sanitize_curse_words}
\title{Sanitize Curse Words in Text}
\usage{
sanitize_curse_words(text, curse_words)
}
\arguments{
\item{text}{A character vector containing the text to be sanitized.}

\item{curse_words}{A character vector containing the curse words to be replaced.}
}
\value{
A character vector with curse words replaced by the placeholder.
}
\description{
This function sanitizes curse words in a given text by replacing them with a designated placeholder.
}
