% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_names.R
\name{sanitize_numbers_in_sentences}
\alias{sanitize_numbers_in_sentences}
\title{Sanitize Numbers in Sentences}
\usage{
sanitize_numbers_in_sentences(sentences_vector)
}
\arguments{
\item{sentences_vector}{A vector of sentences to be sanitized.}
}
\value{
A vector of sanitized sentences.
}
\description{
This function sanitizes occurrences of numbers within a vector of sentences.
}
