% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_names.R
\name{sanitize_topics_in_sentences}
\alias{sanitize_topics_in_sentences}
\title{Sanitize Topics in Sentences}
\usage{
sanitize_topics_in_sentences(sentences_vector, topics_to_sanitize)
}
\arguments{
\item{sentences_vector}{A vector of sentences to be sanitized.}

\item{topics_to_sanitize}{A vector of topics to be anonymized.}
}
\value{
A vector of sanitized sentences.
}
\description{
This function sanitizes occurrences of specified topics within a vector of sentences.
}
