% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_names.R
\name{string_anonymization_functions}
\alias{string_anonymization_functions}
\title{String Anonymization and Sanitization Functions}
\description{
This script provides a collection of functions for anonymizing and sanitizing strings within text data.
These functions are particularly useful for preparing textual data for analysis while ensuring sensitive
information is removed or obfuscated.
}
\section{Functions}{

\itemize{
\item \code{max_typos} - Determines the maximum allowed typos based on word length.
\item \code{anonymize_names} - Sanitizes occurrences of names within a given sentence.
\item \code{anonymize_positions} - Sanitizes occurrences of specified positions within a vector of sentences.
\item \code{sanitize_topics_in_sentences} - Sanitizes occurrences of specified topics within a vector of sentences.
\item \code{sanitize_numbers_in_sentences} - Sanitizes occurrences of numbers within a vector of sentences.
\item \code{sanitize_courses_in_sentences} - Sanitizes occurrences of specified courses within a vector of sentences.
}
}

