% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_coalesce}
\alias{string_coalesce}
\title{Coalesce Multiple Strings}
\usage{
string_coalesce(...)
}
\arguments{
\item{...}{Multiple strings to coalesce.}
}
\value{
A single concatenated string for each row, with NA values ignored.
}
\description{
This function concatenates multiple strings in a row-wise manner, ignoring NA values.
}
