% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_concatenate}
\alias{string_concatenate}
\title{Concatenate Multiple Strings}
\usage{
string_concatenate(...)
}
\arguments{
\item{...}{Multiple strings to concatenate.}
}
\value{
A single concatenated string.
}
\description{
This function concatenates multiple strings into a single string.
}
