% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_drop_before_unsure}
\alias{string_drop_before_unsure}
\title{Drop Text Before "Unsure" in Strings}
\usage{
string_drop_before_unsure(x)
}
\arguments{
\item{x}{A string from which text before "unsure" will be removed.}
}
\value{
A string with text before "unsure" removed.
}
\description{
This function removes text before "unsure" in a string.
}
