% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_drop_columns_containing}
\alias{string_drop_columns_containing}
\title{Drop Columns Containing a Specified String}
\usage{
string_drop_columns_containing(data, string_to_drop)
}
\arguments{
\item{data}{A dataframe from which columns will be dropped.}

\item{string_to_drop}{A string pattern to match column names to drop.}
}
\value{
A dataframe with specified columns dropped.
}
\description{
This function drops columns from a dataframe that contain a specified string in their names.
}
