% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_drop_unsure}
\alias{string_drop_unsure}
\title{Drop "Unsure" From Strings}
\usage{
string_drop_unsure(x)
}
\arguments{
\item{x}{A string from which ", unsure" will be removed.}
}
\value{
A string with ", unsure" removed.
}
\description{
This function removes ", unsure" from a string.
}
