% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_functions}
\alias{string_functions}
\title{String Manipulation Functions for Variable Names, Wordclouds, and Ngrams}
\description{
This collection of functions provides various utilities for handling and manipulating strings. They are primarily
designed for tasks such as renaming variables, cleaning text data, and preparing data for wordclouds and ngram analysis.
}
\section{Functions}{

\itemize{
\item \code{string_concatenate} - Concatenates multiple strings into one string.
\item \code{string_coalesce} - Concatenates multiple strings in a row-wise manner, ignoring NA values.
\item \code{string_rename_columns} - Renames columns in a dataframe based on a replacements list.
\item \code{string_drop_columns_containing} - Drops columns containing a specified string from a dataframe.
\item \code{string_remove_between_parentheses} - Removes text between parentheses from a string.
\item \code{string_replace_multiple} - Replaces multiple patterns in a text column based on a replacements list.
\item \code{string_drop_unsure} - Removes ", unsure" from a string.
\item \code{string_drop_before_unsure} - Removes text before "unsure" in a string.
\item \code{string_remove_unwanted} - Removes unwanted symbols from a string.
\item \code{string_tokenize_data} - Tokenizes text data for wordcloud and ngram analysis.
}
}

