% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_remove_unwanted}
\alias{string_remove_unwanted}
\title{Remove Unwanted Symbols From Strings}
\usage{
string_remove_unwanted(input_string)
}
\arguments{
\item{input_string}{A string from which unwanted symbols will be removed.}
}
\value{
A cleaned string with unwanted symbols removed.
}
\description{
This function removes unwanted symbols from a string.
}
