% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_rename_columns}
\alias{string_rename_columns}
\title{Rename Columns in a Dataframe}
\usage{
string_rename_columns(data, replacements)
}
\arguments{
\item{data}{A dataframe whose columns are to be renamed.}

\item{replacements}{A named list of replacements where names are the patterns to replace and values are the replacements.}
}
\value{
A dataframe with renamed columns.
}
\description{
This function renames columns in a dataframe based on a replacements list.
}
