% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R
\name{string_tokenize_data}
\alias{string_tokenize_data}
\title{Tokenize Text Data for Wordclouds and Ngrams}
\usage{
string_tokenize_data(data, text_column, ngram = 2)
}
\arguments{
\item{data}{A dataframe containing the text data.}

\item{text_column}{The name of the text column to be tokenized.}

\item{ngram}{The number of words in each ngram. Default is 2.}
}
\value{
A list containing wordcloud data and ngram data.
}
\description{
This function tokenizes text data, removes stop words, and prepares data for wordcloud and ngram analysis.
}
