% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_basic.R
\name{theme_basic}
\alias{theme_basic}
\title{Create my default theme, borrowing from cowplot - https://github.com/wilkelab/cowplot/blob/1f35f385b0631c74daecaf6366725a8573242bde/R/themes.R#L11
library(ggplot2)
library(ggtext)
library(cowplot)}
\usage{
theme_basic()
}
\value{
A ggplot2
}
\description{
Create my default theme, borrowing from cowplot - https://github.com/wilkelab/cowplot/blob/1f35f385b0631c74daecaf6366725a8573242bde/R/themes.R#L11
library(ggplot2)
library(ggtext)
library(cowplot)
}
