% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary.R
\name{make_summary}
\alias{make_summary}
\title{This function generates a summary table with various statistics for the specified grouping variables and dependent variable.}
\usage{
make_summary(data, dv, grouping1, grouping2, grouping3)
}
\arguments{
\item{data}{The targeted dataframe}

\item{dv}{The name of the variable we want to summarize}

\item{grouping1}{One grouping variable}

\item{grouping2}{A second grouping variable}

\item{grouping3}{A third grouping variable}
}
\value{
A dataframe with summary statistics
}
\description{
This function generates a summary table with various statistics for the specified grouping variables and dependent variable.
}
